Split SDE tests in half, to try and avoid GitHub runner issues?
